Introduce `DefaultLocale` Value Object
We need a 'default' locale in many situations. It makes sense to locate this default in the main I18n library, but rather than put a string in DI, we have a simple, stringable value object.

The default is guaranteed by falling back on `Locale::getDefault()` in the factory for this value object.

Signed-off-by: George Steel <george@net-glue.co.uk>